Add a wasi-crypto proposal for signatures and a PoC implementation
The implementation could be better and is missing the wasm<->host glue
to be usable from WebAssembly, but it can help evaluate if the API is
usable and what are its limitations.

Only signatures are implemented. But signatures are quite complicated
to design an API for, so this is probably a good starting point for
the overall API design.

An "operation" object must be created. For signatures, this is a
`signature_op` object, instantiated with an algorithm name.

Different key types are used for each type of operation, and a key
is tied to a context, i.e. an algorithm and a set of parameters.

Additional functions to tweak the context can be added without
breaking the current API.

A `signature_keypair_builder` object, tied to a context, can create a
key pair:

- Either by generating a new one, if supported by the platform for the
given algorithm
- Using a key identifier for HSM-like scenarios (we may need a
version, too, but that can be discussed later)
- By importing data using a specified encoding.

Key pairs are opaque handles. The host may not allow secret keys to be
exported and/or imported.

Having an intermediary "builder" object allows applications to set up
additional parameters required for key generation, importation (ex:
password) or reference (ex: HSM credentials).

If allowed, key pairs can be exported, using a specified encoding.

We should define what format MUST and MAY be supported by an
implementation for common algorithms.

In particular, we may require PKCS8 to be supported whenever the
ASN.1 encoding has been defined for an algorithm.

A public key object can obviously be created from a key pair, and
public keys can be imported and exported as well.

The signature process itself goes through the creation of a
`signature_state` object, on which an `update` function can add to the
message being signed.

At any point in time, the `sign` function of `signature_state` can be
called to compute a signature of all the data received so far.
The state is not reset, in order to allow incremental signatures.

Non-deterministic signature systems (ECDSA, RSA padding) do not
require applications to specify a nonce. In order to avoid
catastrophic security failures, the `wasi-crypto` implementation is
responsible for securely choosing the nonce instead of delegating this
to applications.

Signature verification works in a similar way. Data is received via
a `signature_verification_state` object that is tied to a public key,
and allows a signature to be verified against the data received so far.

We use quite a lot of different object types here. While it makes the
API slightly more verbose than if more generic types had been used, it
can make static analysis far more useful.